rename Button to HKButton
