+++ src/index.ts
@@ -1,2 +1,2 @@
-export { default as Button } from './Button'
+export { default as HKButton } from './HKButton'
 export { default as Modal } from './Modal'
